requirements.txt
### **requirements.txt**
```txt
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
numpy>=1.25.0
tqdm>=4.65.0